% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition.R
\name{addition}
\alias{addition}
\title{Addition}
\usage{
addition(x, y)
}
\arguments{
\item{x}{one value}

\item{y}{next value}
}
\value{
an integer
}
\description{
Basic addition
}
\details{

}
\examples{
add <- addition(1,2)

}
