% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.R
\name{divide}
\alias{divide}
\title{Divide}
\usage{
divide(x, y)
}
\arguments{
\item{x}{one value}

\item{y}{next value}
}
\value{
an integer
}
\description{
Basic division
}
\details{

}
\examples{
div <- divide(10,2)

}
