% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanicFunction.R
\name{gpFunction}
\alias{gpFunction}
\title{Illustration of gpFunction}
\usage{
gpFunction(Sex, Class, Age, Family, Embarked, Fare, model = NULL)
}
\arguments{
\item{Sex}{male, female: include quotation marks (\code{"male"} or \code{"female"})}

\item{Class}{First, Second, Third,; include quotation marks 
(\code{"First"} or \code{"Second"} or \code{"Third"})}

\item{Age}{real number}

\item{Family}{integer}

\item{Embarked}{C,Q,S; include quotation marks (\code{"C"} or \code{"Q"} or \code{"S"})}

\item{Fare}{real number}

\item{model}{By default this is the model given by stepwise regression using the starting model 3 
(Model including all three-way interaction terms and lower terms, plus some polynomial terms). Can be 
changed.}
}
\value{
The probability of someone with those characteristics surviving is returned.
}
\description{
Returns the probability of the given person surviving or not.
}
\examples{
Test1 <- gpFunction("male", "Third", 20, 0, "S", 8.05)
Test2 <- gpFunction("female", "Third", 21, 0, "Q", 7.8792)
Test3 <- gpFunction("male", "First", 40, 0, "C", 27.7208)
Test4 <- gpFunction("female", "Second", 3, 3, "C", 41.5792)

}

