% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply}
\alias{multiply}
\title{Multiply}
\usage{
multiply(x, y)
}
\arguments{
\item{x}{one value}

\item{y}{next value}
}
\value{
an integer
}
\description{
Basic division
}
\details{

}
\examples{
mutli <- multiply(5,5)

}
