% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtraction.R
\name{subtraction}
\alias{subtraction}
\title{Subtraction}
\usage{
subtraction(x, y)
}
\arguments{
\item{x}{one value}

\item{y}{next value}
}
\value{
an integer
}
\description{
Basic subtraction
}
\details{

}
\examples{
sub <- subtraction(5,2)

}
